requirements.txt
pymongo==4.10.1
python-dotenv==1.0.1
pandas==2.2.3
jsonschema==4.21.1
boto3==1.35.84
numpy==1.26.4
botocore==1.35.84 